Fix Mode2 lux calculation, add I2C error check, fix example pin comment
- GetLightIntensity(): divide by 2 for H-Resolution Mode2 variants
  (k_DevModeContHighRes2 / k_DevModeOneTimeHighRes2) per datasheet p.11;
  without this fix Mode2 returned double the correct lux value (upstream issue #12)
- GetLightIntensity(): change return type uint16_t -> float to preserve
  sub-lux precision lost by truncating the / 1.2 result
- GetLightIntensity(): check Wire.requestFrom() byte count; return -1.0
  instead of reading garbage bytes when sensor is absent or bus is stuck
- BH1750FVI_Simple.ino: correct ADDR pin comment from 'RX' to 'D7
  (GPIO13, pin 13)' to match ADDRESSPIN = 13 in the code
- Update both example sketches to use float lux to match new return type

+++ src/BH1750FVI.h
@@ -83,9 +83,9 @@ class BH1750FVI {
 
     /*
      * Get the measured light intensity from the device
-     * @return The light intensity measured by the device in 'lux'
+     * @return The light intensity measured by the device in 'lux', or -1.0 on I2C error
      */
-    uint16_t GetLightIntensity(void);
+    float GetLightIntensity(void);
     
     /*
      * Set the mode
